Создана страница MainPage, LoginPage, настроены 2 роута
