+++ package.json
@@ -14,7 +14,7 @@
     "react-dom": "^17.0.1",
     "react-intersection-observer": "^8.32.5",
     "react-redux": "^7.2.9",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
