+++ src/components/app/app.jsx
@@ -1,55 +1,25 @@
-import { useEffect } from "react";
-import styles from "./app.module.css";
-import AppHeader from "../app-header/app-header";
-import BurgerIngredients from "../burger-ingredients/burger-ingredients";
-import BurgerConstructor from "../burger-constructor/burger-constructor";
-import Modal from "../modal/modal";
-import OrderDetails from "../order-details/order-details";
-import IngredientDetails from "../ingredient-details/ingredient-details";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import { DndProvider } from "react-dnd";
-import { useDispatch, useSelector } from "react-redux";
-import { loadIngredients } from "../../services/ingredients";
+import MainPage from "../pages/main";
+import LoginPage from "../pages/login";
+
+// import {ProtectedRouteElement} from "./components/protected-route";
+// import {ProvideAuth} from "./services/auth";
+
+{
+  /* <Route path="/" element={<ProtectedRouteElement element={<HomePage />}/>} /> */
+}
 
 function App() {
-  const dispatch = useDispatch();
-  const modalData = useSelector((state) => state.modalData);
-
-  const { loading, error } = useSelector((state) => state.ingredientsData);
-
-  useEffect(() => {
-    dispatch(loadIngredients());
-  }, []);
-
   return (
     <DndProvider backend={HTML5Backend}>
-      <AppHeader />
-      <main className={styles.main}>
-        {!error && (
-          <>
-            <BurgerIngredients />
-            <BurgerConstructor />
-          </>
-        )}
-        {loading && (
-          <p className="text text_type_main-large">Загрузка данных...</p>
-        )}
-        {error && (
-          <p className="text text_type_main-large">
-            Ошибка, попробуйте позже ({error.message})
-          </p>
-        )}
-      </main>
-      <div className={styles.modals_container}>
-        {modalData.active && (
-          <Modal>
-            {modalData.type === "order" && <OrderDetails />}
-            {modalData.type === "ingredient" && (
-              <IngredientDetails ingredient={modalData.ingredient} />
-            )}
-          </Modal>
-        )}
-      </div>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<MainPage />} />
+          <Route path="/login" element={<LoginPage />} />
+        </Routes>
+      </BrowserRouter>
     </DndProvider>
   );
 }
