+++ src/components/pages/main.module.css
@@ -1,0 +1,9 @@
+.main {
+    display: flex;
+    justify-content: center;
+    gap: 40px;
+  }
+  
+  .modals_container {
+    overflow: hidden;
+  }
